requirements.txt
aiogram>=3.7.0
telethon>=1.36.0
aiosqlite>=0.20.0
aiohttp>=3.9.5
python-dotenv>=1.0.1
